restructured modules and packages
