+++ tests/test_trainer.py
